[#69984] vo 변수 수정, UserDao부분 쿼리추가

+++ src/com/zum/pilot/vo/UserVo.java
@@ -1,16 +1,17 @@
 package com.zum.pilot.vo;
 
 public class UserVo {
-	private int id;
-	private String email;
-	private String name;
-	private String password;
-	private String create_time;
-	private String update_time;
-	public int getId() {
+	private Long id;	// 유저 id
+	private String email;	// 유저 이메일
+	private String name;	// 유저 닉네임
+	private String password;	// 유저 비밀번호
+	private String create_time;	// 가입일
+	private String update_time;	// 회원수정일
+	
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getEmail() {
@@ -43,6 +44,4 @@ public class UserVo {
 	public void setUpdate_time(String update_time) {
 		this.update_time = update_time;
 	}
-	
-	
 }
